+++ js/const.js
@@ -1,0 +1,23 @@
+const TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'],
+  TIMES = ['12:00', '13:00', '14:00'],
+  FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'],
+  PHOTOS = [
+    'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+    'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+    'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+  ],
+  // OFFERS_COUNT = 10,
+  PRICE_START = 100,
+  PRICE_END = 10000,
+  ROOMS_START = 1,
+  ROOMS_END = 10,
+  GUESTS_START = 1,
+  GUESTS_END = 20,
+  LAT_START = 35.65000,
+  LAT_END = 35.70000,
+  LAT_DIGITS = 5,
+  LNG_START = 139.70000,
+  LNG_END = 139.80000,
+  LNG_DIGITS = 5;
+
+export {TYPES,TIMES,FEATURES,PHOTOS,PRICE_START,PRICE_END,ROOMS_START,ROOMS_END,GUESTS_START,GUESTS_END,LAT_START,LAT_END,LAT_DIGITS,LNG_START,LNG_END,LNG_DIGITS};
